bpf: enable -nostdinc and other stricter compilation options
We want to have our code as close to usual kernel conventions as
possible to ease review, so add a few more switches for clang
where we emit useful warnings in our code if it's not the case,
that is, -std=gnu89 and -Wdeclaration-after-statement.

We also use quite a number of shadow declarations where they are
really useless; I'm not sure whether clang ends up allocating new
stack space for them; in any case lets add -Wshadow given the
current locations are fixed now.

Also, given our headers are all self-contained, add -nostdinc to
fully /guarantee/ that clang does not search the standard include
paths anymore.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>